+++ CHANGELOG.md
@@ -10,10 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add logic to ignore some Attack speed update. Prevent to update swing timer on UNIT_ATTACK_SPEED when Druid shapeshift.
 - Add spell id for swing spell reset for Warlok, Mage and Priest Shoot ability.
 - Add swing timer pause logic (Warrior Slam mechanic).
-- Add lib MAJOR MINOR version check.
+- Add LibStub version managment.
 
 ### Changed
-- Rename SwingTimerAPI global var to SwingTimerLib.
 - Init the Lib variable after ADDON_LOADED event.
 
 ### Fixed